[4] move register page to auth feature
