+++ frontend/src/features/auth/Register/register.routes.tsx
@@ -1,6 +1,6 @@
 import { createRoute } from "@tanstack/react-router";
 import Register from "./Register.page";
-import { rootRoute } from "../../routes/router";
+import { rootRoute } from "../../../routes/router";
 
 export const registerRoute = createRoute({
   getParentRoute: () => rootRoute,
